Feat/http version (#94)
Fix #87, Fix #50 
* :sparkles: Implement HTTP version check
- Add Version class
- Add 505 error page

* Keep-Alive is default for HTTP/1.1 or later
- Connection is closed by default for HTTP/1.0 or earlier

* Fix #50 - Chunked encoding is not supported in HTTP/1.0
A server or client that receives an HTTP/1.0 message containing a Transfer-Encoding header field MUST treat the message as if the framing is faulty, even if a Content-Length is present, and close the connection after processing the message. The message sender might have retained a portion of the message, in buffer, that could be misinterpreted by further use of the connection.